Some adjustments
There is now a cutout in Jig1 for the JST connector, I've adjusted the distance on Jig2 so that the halls get as close to the magnets as possible, and Jig 3 is now its own jig, also getting the halls as close as possible to the magnets. the readme has been adjusted to reflect this.